namespace interfaces with underscore
need an unambiguous interface name to avoid confusion or misuse in the future, for example with karma relations. Since interface names must be unique, it would be impossible for alertmanager to have two 'alertmanager' interface names, one for prometheus and one for karma, even if they are under different relation names.