Integration test for templated notification by receiver (#74)
* Added support for setting alerts in Client

This commit adds support for setting alerts in the Alertmanager
client object. This support is needed to create integration tests
which involve triggering alerts and checking that these alerts
are correctly forwarded to the receiver.

* Fix double quoted template file path

This commit fixes a bug that results in a failure of alertmanager
templates configuration option from working as expected. The
commit also fixes a corresponding unit tests.

* An integration test altermanger templates

This commit adds and integration test to check that the alertmanger
templates configuration option works as expected. The test checks
that a fake alert receiver does get notifications that are customized
using a template.

Closes #40

* Added unit test for alertmanager clean set alerts

This commit adds a unit tests for alert setting api of
alertmanager client.

* Added type hints to Alertmanager client

This commit adds a few type hints to the new methods of alertmanager
client.

Co-authored-by: Leon <82407168+sed-i@users.noreply.github.com>