Integration test for persistence of silences (#78)
* Enhance alertmanager client with silences API

This commit enhances the Alertmanager client object to add support
for more methods such as setting, getting and deleting silences,
in addition to getting current list of alerts.

* An integration test for persistence of silences

This commit implements and integration test to check for the
persistence of silences. The test deploys Alertmanager from
charmhub, then sets a silence, then upgrades Alertmanger using
a locally built charm, and finally checks if the silence still
persists.

* Use kubectl to send a sigterm signal to Alertmanager

This commit uses kubectl to send a SIGTERM signal to alertmanager
during the integration test for persistence of silences. This
is temporarily necessary because this bug
https://github.com/canonical/pebble/issues/122 .

* Added unit tests for enhanced alertmanager client

This commit adds a few more unit tests for the recent
enhancements made to the alertmanager client.